requirements.txt
streamlit
gspread
oauth2client
pydrive
pandas
Pillow
folium
streamlit_folium